Code cleanup
Cleaning up various warnings from compiling this on OSX and
trying to get it to work. It keeps segfaulting due to issues
with a linked list in the hash table not working properly.
Cleaning up the warnings in attempt to see if anything fixes it.
And because well, fixing warnings is never bad.