Misc warning fixes and memory segfault fix
Fixed up a bunch of warnings and worked on getting the port loading
to not cause a segfault. Issue was a malloc that was not making
enough space in the hashtable. That code has worked forever,
it's possible it was making use of a more liberal compiler that
was assuming the correct size based on sizeof(struct list *)