Minor bigbang fixes, moved save functions out
Still going through valgrind and trying to make changes to bigbang
so that it doesn't keep mucking things up. To facilitate this further
I removed the save functions from severactions.c so that they would
be easier to deal with on their own. Ultimately I want to use the
same saving functions everywhere.