Bigbang struct init fixes and other misc
Fixed up the structs in bigbang to be initalized using designated
initalizers. This cleared up a whole slew of issues that valgrind
was finding. Also, the constellation name generator was running
out of names too quick and there wasn't a loop check to ensure that
didn't happen. I fixed that. This seems to have fixed the issues
bigbang was having with size. However, there are still issues
when using node creation (per valgrind). That's an issue for
another day though.