Config file fix and more autoconf fixes
There was a discrepancy in what config file bigbang was loading.
Also compile was a symbolic link which really shouldn't be included.
There's still some oddities with the automake autoheader stuff that I
need to look into. Back in the day it just used to be running
configure and make. Most of the issue I run into them now is misaligned
automake versions.